refactor model_loader
